style: automate system props split ( wrapper / component / system )
