+++ package.json
@@ -101,6 +101,7 @@
     "downshift": "=3.2.10",
     "emoji-mart": "=2.11.1",
     "lodash.concat": "^4.5.0",
+    "lodash.difference": "^4.5.0",
     "lodash.get": "^4.4.2",
     "lodash.isequal": "=4.5.0",
     "lodash.kebabcase": "=4.1.1",
