+++ src/components/Select/doc.mdx
@@ -29,16 +29,7 @@ It is based on the [downshift](https://github.com/downshift-js/downshift) librar
     ]
     const [value, setValue] = useState()
     const handleChange = e => setValue(e.target.value)
-    return (
-      <Select
-        options={ITEMS}
-        name="default"
-        onChange={handleChange}
-        value={value}
-        mt={20}
-        borderRadius={15}
-      />
-    )
+    return <Select options={ITEMS} name="default" onChange={handleChange} value={value} />
   }}
 </Playground>
