+++ src/usage-xstyled.mdx
@@ -93,11 +93,13 @@ We split these props at the container level so that…
                 options={networks}
                 label="Social network"
                 placeholder="Choose network"
-                mt={120}
-                mx="25%"
+                mt={50}
+                width={150}
                 height="auto"
-                padding={20}
-                borderRadius={50}
+                borderColor="papayaWhip"
+                borderWidth={2}
+                borderRadius={15}
+                padding={50}
               />
             </form>
           )
@@ -124,13 +126,13 @@ We split these props at the container level so that…
           { value: 'madeleine', label: 'Madeleine (made with farm-fresh eggs)' },
           { value: 'pumpkin', label: 'Pumpkin' }
         ]}
-        mt={120}
-        mr="50%"
+        mt={50}
+        width={150}
         height="auto"
-        px={25}
-        py={35}
-        borderRadius={0}
-        borderWidth={10}
+        borderColor="papayaWhip"
+        borderWidth={2}
+        borderRadius={15}
+        padding={50}
       />
     </>
   )}
