+++ src/utils/index.js
@@ -5,7 +5,7 @@ export * from './font'
 export * from './getMax'
 export * from './hexToRGB'
 export * from './propTypes'
+export * from './system'
 export * from './text'
-export * from './utils'
 export * from './validations'
 export * from './variants'
